+++ ThirdPartyAdapters/mytarget/mytarget/src/main/java/com/google/ads/mediation/mytarget/MyTargetAdapterUtils.kt
@@ -13,6 +13,7 @@
 
 package com.google.ads.mediation.mytarget
 
+import com.google.android.gms.ads.AgeRestrictedTreatment
 import com.google.android.gms.ads.RequestConfiguration
 import com.my.target.common.MyTargetPrivacy
 
@@ -25,10 +26,13 @@ object MyTargetAdapterUtils {
   fun configureMyTargetPrivacy(requestConfiguration: RequestConfiguration) {
     val isChildDirected = requestConfiguration.tagForChildDirectedTreatment
     val isUnderAgeOfConsent = requestConfiguration.tagForUnderAgeOfConsent
+    val ageRestrictedTreatment = requestConfiguration.ageRestrictedTreatment
 
     if (
       isChildDirected == RequestConfiguration.TAG_FOR_CHILD_DIRECTED_TREATMENT_TRUE ||
-        isUnderAgeOfConsent == RequestConfiguration.TAG_FOR_UNDER_AGE_OF_CONSENT_TRUE
+        isUnderAgeOfConsent == RequestConfiguration.TAG_FOR_UNDER_AGE_OF_CONSENT_TRUE ||
+        ageRestrictedTreatment == AgeRestrictedTreatment.CHILD ||
+        ageRestrictedTreatment == AgeRestrictedTreatment.TEEN
     ) {
       MyTargetPrivacy.setUserAgeRestricted(true)
     } else if (