+++ ThirdPartyAdapters/mytarget/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## myTarget Android Mediation Adapter Changelog
 
 #### Next Version
+- Maps `AgeRestrictedTreatment` to myTarget's age restriction settings.
 - Updated interstitial ad implementation to use updated mediation adapter APIs.
 
 #### Version 5.47.1.0 (In progress)